chore: split harbor out a bit, and continue on errors in a single namespace to prevent blocking
